% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{create.simulations}
\alias{create.simulations}
\title{Creates simulations of group sequential trial data and
calculates test statistic based on user-defined stat function.}
\usage{
create.simulations(n_sims, n_k, stat.func, data.generator, ...)
}
\arguments{
\item{n_sims}{Number of simulations}

\item{n_k}{Sample size per stage}

\item{stat.func}{Function (or list of functions)
to calculate the test statistic based on the generated data, at each
observed information rates. If this is a list
of functions, it needs to be the same length as the sequence of boundaries.}

\item{data.generator}{Function to generate data. The output of this function
should be a matrix, which is the input for stat.func.
First argument of this function needs to be n, the sample size, and the function
needs to be vectorized.}

\item{...}{Additional arguments to be passed to \code{data.generator}}
}
\description{
Creates simulations of group sequential trial data and
calculates test statistic based on user-defined stat function.
}
\examples{
create.simulations(10, n_k=c(10, 10, 10),
                   stat.func=list(mean, sum, mean),
                   data.generator=function(n) as.matrix(rnorm(n)))
}
