% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{get.power}
\alias{get.power}
\title{Calculate the type I error based on a sequence of boundaries.
Inputs include (1) trial parameters, (2) data generating procedure parameters,
and (3) a function to compute the test statistic based on observations and stage.}
\usage{
get.power(u, simulations, previous = NULL)
}
\arguments{
\item{u}{Constant boundary}

\item{simulations}{Output from create.simulations function.
Should be long by simulation and wide by stage.}

\item{previous}{A vector of previous conditions.}
}
\description{
Calculate the type I error based on a sequence of boundaries.
Inputs include (1) trial parameters, (2) data generating procedure parameters,
and (3) a function to compute the test statistic based on observations and stage.
}
\examples{
simulations <- matrix(rnorm(100), nrow=100)
get.power(2.7, simulations=simulations)

previous <- rep(TRUE, 10)
}
