% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{obf.spend}
\alias{obf.spend}
\title{O'Brien-Fleming approximation, alpha spending function}
\usage{
obf.spend(a)
}
\arguments{
\item{a}{Alpha level}
}
\description{
O'Brien-Fleming approximation, alpha spending function
}
