% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{pocock.spend}
\alias{pocock.spend}
\title{Pocock approximation, alpha spending function}
\usage{
pocock.spend(a)
}
\arguments{
\item{a}{Alpha level}
}
\description{
Pocock approximation, alpha spending function
}
