% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{solve.boundary}
\alias{solve.boundary}
\title{Root solve for a particular alpha level.}
\usage{
\method{solve}{boundary}(power, simulations, previous = NULL)
}
\arguments{
\item{power}{The type I error (or power generally) to solve for}

\item{simulations}{Output from create.simulations function.}
}
\description{
Root solve for a particular alpha level.
}
