% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{validate.trial.params}
\alias{validate.trial.params}
\title{Validate parameters to use in the type I error functions.
Don't want to include this in the function itself because
it would be run multiple times when solving for the boundaries.}
\usage{
validate.trial.params(n_k, stat.func, data.generator, ...)
}
\arguments{
\item{n_k}{Sample size or sequence of sample sizes for each stage.}

\item{stat.func}{Function (or list of functions)
to calculate the test statistic based on the generated data. If this is a list
of functions, it needs to be the same length as the sequence of boundaries.}

\item{data.generator}{Function to generate data. The output of this function
should be a matrix, which is the input for stat.func.
First argument of this function needs to be n, the sample size, and the function
needs to be vectorized.}

\item{...}{Additional arguments to the \code{data.generator} function}
}
\description{
Validate parameters to use in the type I error functions.
Don't want to include this in the function itself because
it would be run multiple times when solving for the boundaries.
}
\examples{
validate.trial.params(n_k=rep(10, 5), stat.func=mean, data.generator=rnorm)

}
