% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{get.boundaries.aspend}
\alias{get.boundaries.aspend}
\title{Derive boundaries with alpha-spending.}
\usage{
get.boundaries.aspend(
  a.func,
  rates,
  u_k = NULL,
  rho = 1,
  change = 0,
  algorithm = Miwa(steps = 1000)
)
}
\arguments{
\item{a.func}{A continuous, monotonic increasing function of t
where a.func(0) = 0 and a.func(1) = a
where a is the type I error desired}

\item{rates}{A vector of information rates (between 0 and 1)}

\item{u_k}{An optional matrix of previous boundaries}

\item{rho}{Fraction of variance explained by fitting ANCOVA.}

\item{change}{A vector indicating which stages use ANOVA v. ANCOVA.}
}
\description{
Get the boundaries required for a particular alpha-spending function
and an observed information rate out of a total sample size.
}
\details{
Your data generator in this case needs to be the data generation
under the null hypothesis in order for this function to work correctly
with alpha-spending.
}
\examples{
# Information rates
t <- 1:4/4

# Approximate Pocock boundaries, with R^2 = 0.5, ANCOVA at last stage
get.boundaries.aspend(a.func=pocock.spend(0.05), rates=t)
get.boundaries.aspend(a.func=pocock.spend(0.05), rates=t,
                      rho=sqrt(0.5), change=c(0, 0, 0, 1))
}
